+++ src/app/bets/bets-list/bets-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { AuthService } from 'src/app/core/services/auth.service';
 import { BetsListService } from 'src/app/core/services/bets-list.service';
 import { Subscription } from 'rxjs';
+import { Bet } from 'src/app/core/models/bet.model';
 
 @Component({
   selector: 'app-bets-list',
@@ -15,15 +16,38 @@ export class BetsListComponent implements OnInit {
   userIsAuthenticated = false;
   userId: string;
   private authStatusSub: Subscription;
-  
-  ELEMENT_DATA: Bet[] = [];
+
+  bet: Bet = {
+    id: 0,
+    fechaCreacion: new Date(),
+    fechaCierre: new Date(),
+    firstPrice: 0,
+    secondPrice: 0,
+    thirdPrice: 0,
+    fare: 0,
+    open: true
+  }
+
+  ELEMENT_DATA: Bet[] = [
+    {
+      id: 1, fechaCreacion: new Date(), fechaCierre: new Date(), firstPrice: 10,
+      secondPrice: 9, thirdPrice: 8, fare: 8, open: true
+    },
+    {
+      id: 2, fechaCreacion: new Date(), fechaCierre: new Date(), firstPrice: 11,
+      secondPrice: 10, thirdPrice: 8, fare: 8, open: false
+    },
+    {
+      id: 3, fechaCreacion: new Date(), fechaCierre: new Date(), firstPrice: 10,
+      secondPrice: 9, thirdPrice: 8, fare: 8, open: true
+    }];
 
   displayedColumns: string[] = ['id', 'fechaCreacion', 'fechaCierre', 'Premio Mayor',
-  'Segundo Premio','Tercer Premio','precio Boleta','Estado'];
+    'Segundo Premio', 'Tercer Premio', 'precio Boleta', 'Estado', 'Eliminar'];
   dataSource = this.ELEMENT_DATA;
 
   constructor(private authService: AuthService,
-              public betService: BetsListService) { }
+    public betService: BetsListService) { }
 
   ngOnInit() {
     this.isLoading = true;
@@ -36,19 +60,14 @@ export class BetsListComponent implements OnInit {
       });
   }
 
-  cargarBets(){
-    this.betService.cargarBets().subscribe( bets => this.ELEMENT_DATA = bets);
+  cargarBets() {
+    this.betService.cargarBets().subscribe(bets => this.ELEMENT_DATA = bets);
+  }
+
+  delete() {
+    this.betService.deleteBets(this.bet)
   }
 }
 
-export interface Bet {
-  id: number,
-  fechaCreacion: Date,
-  fechaCierre: Date,
-  firstPrice: number,
-  secondPrice: number,
-  thirdPrice: number,
-  fare: number,
-  open: Boolean
-}
+
