+++ src/app/core/models/bet.model.ts
@@ -1,0 +1,10 @@
+export interface Bet {
+  id: number,
+  fechaCreacion: Date,
+  fechaCierre: Date,
+  firstPrice: number,
+  secondPrice: number,
+  thirdPrice: number,
+  fare: number,
+  open: Boolean
+}
