+++ src/app/core/services/bets-list.service.ts
@@ -13,6 +13,12 @@ export class BetsListService {
   cargarBets() : any{
     return this.http.get( BACKEND_URL );
   }
+
+  deleteBets( bet: Bet){
+    return this.http.delete(BACKEND_URL + `/${bet.id}`);
+
+
+  }
 }
 
 export interface Bet {
